Change name and move config implementation class to its own file
